Implemented the front end of pagination
After a while I finally got back to pagination:
I took an old function I did once in PHP and translated it into
JavaScript for use with AJAX. Coupled with preparations I did last week
in back end, it was just a matter of choosing good elements for the
pagination and passing the current page back to server, and then using
the results as before.

The pagination I had done a few months ago was made for use with PHP
Foundation classes, so I had to adapt that as well for use with
Bootstrap.